Atualizei o Ex07

+++ Mundo-01/Ex07.py
@@ -1,3 +1,3 @@
-nota1 = int(input('Digite a nota da Prova 01: '))
-nota2 = int(input('Digite a nota da Prova 02: '))
+nota1 = float(input('Digite a nota da Prova 01: '))
+nota2 = float(input('Digite a nota da Prova 02: '))
 print(f'A média do aluno é {(nota1+nota2)/2:.2f}')
